backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib==1.7.4
python-jwt==4.1.0
pydantic==2.6.3